+++ src/compiler/llvm_codegen_function.c
@@ -663,7 +663,7 @@ void llvm_emit_dynamic_functions(GenContext *c, Decl **funcs)
 
 void llvm_emit_function_decl(GenContext *c, Decl *decl)
 {
-	ASSERT(decl->decl_kind == DECL_FUNC);
+	ASSERT_SPAN(decl, decl->decl_kind == DECL_FUNC);
 	// Resolve function backend type for function.
 	decl_append_links_to_global(decl);
 	LLVMValueRef function = llvm_get_ref(c, decl);